read raw Wacom Bamboo CTH-470 data via libusb
works on macOS; much better than working with the macOS facilities that are
poorly documented and verbose as fuck (not to mention hours wasted on the HID
Manager stuff that doesn't even know how to talk to the Wacom, and the official
"drivers" that grabs exclusive access to the device). a bit hackish, partly
based on Wacom driver in Linux, partly reverse engineered.